ai_estimation/requirements.txt
pyside6>=6.6.1
openai>=1.10.0
pandas>=2.2.0
openpyxl>=3.1.2
pytest>=8.0.0
python-dotenv>=1.0.0 